# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Game Evolution",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/turns": {
            "get": {
                "description": "Возвращает все ходы с фильтрацией по статусу и дате формирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ходы"
                ],
                "summary": "Получить все ходы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "статус ходы",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "начальная дата формирования",
                        "name": "formation_date_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "конечная дата формирвания",
                        "name": "formation_date_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllTurnsResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Позволяет изменить тип чернового ходы и возвращает обновлённые данные",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ходы"
                ],
                "summary": "Указать тип ходы",
                "parameters": [
                    {
                        "description": "Тип ходы",
                        "name": "turn_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SwaggerUpdateTurnRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.TurnOutput"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет черновое карту",
                "tags": [
                    "Ходы"
                ],
                "summary": "Удалить черновое карту",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/turns/delete_card/{id}": {
            "delete": {
                "description": "Удалить карты из черновово ходы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ходы"
                ],
                "summary": "Удалить карты из черновово ходы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id карты",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllCardsResponse"
                        }
                    }
                }
            }
        },
...
        "/api/turns/{id}": {
            "get": {
                "description": "Возвращает подробную информацию об ходы и его типе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ходы"
                ],
                "summary": "Получить одно карту",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id ходы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.TurnResponse"
                        }
                    }
                }
            }
        },
        "/api/turns/{id}/moderator_confirm": {
            "put": {
                "description": "Подтвердить или отменить карту модератором",
                "tags": [
                    "Ходы"
                ],
                "summary": "Подтвердить карту",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id ходы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "подтвердить",
                        "name": "confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.TurnOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.SwaggerUpdateTurnRequest": {
            "type": "object",
            "properties": {
                "turn_type": {
                    "type": "string"
                }
            }
        },
        "ds.Card": {
            "type": "object",
            "required": [
                "description",
                "name",
                "needfood",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 200
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "needfood": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "maxLength": 50
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "schemes.AllCardsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Card"
                    }
                }
            }
        },
        "schemes.AllTurnsResponse": {
            "type": "object",
            "properties": {
                "turns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemes.TurnOutput"
                    }
                }
            }
        },
...
        "schemes.TurnOutput": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "moderator": {
                    "type": "string"
                },
                "sending_status": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "takefood": {
                    "type": "integer"
                },
                "turn_type": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.TurnResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Card"
                    }
                },
                "turn": {
                    "$ref": "#/definitions/schemes.TurnOutput"
                }
            }
        },
...
    }
}